Separate serial code from main.py
