+++ robot/urov_pi/main.py
@@ -1,45 +1,11 @@
-import serial
+import socket   #for sockets connection to RPi
+import sys  #for exit
+from time import sleep #for waiting
 
 import packet
 
 __author__ = 'Spencer Shaw'
 
-class packet:
+serial_coms = new SerialConnection()
 
-    def __init__(self, cmd, value1, value2, seqnum):
-        self.cmd = cmd
-        self.value1 = value1
-        self.value2 = value2
-        self.seqnum_chksum = (seqnum << 4) + calc_checksum(cmd, value1, value2, seqnum)
-
-outbound = serial.Serial(
-    port=port,
-    baudrate=9600,
-    parity=serial.PARITY_NONE,   # parity is error checking, odd means the message should have an odd number of 1 bits
-    stopbits=serial.STOPBITS_ONE,
-    bytesize=serial.EIGHTBITS,   # eight bits of information per pulse/packet
-    timeout=0.1
-)
-
-class packet:
-
-    def __init__(self, cmd, value1, value2, seqnum):
-        self.cmd = cmd
-        self.value1 = value1
-        self.value2 = value2
-        self.seqnum_chksum = (seqnum << 4) + calc_checksum(cmd, value1, value2, seqnum)
-        
-def establishContact():
-    sendPacket(new packet(0x00, 0x00, 0x00, 0x00))
-   
-
-def sendPacket(packet p):
-    outbound.write(p.cmd)
-    outbound.write(p.value1)
-    outbound.write(p.value2)
-    outbound.write(p.seqnum_chksum)
-    
-def calc_chksum(cmd, value1, value2, seqnum):
-    return (cmd + (value1 * 3) + (value2 * 5) + (seqnum_chksum * 7)) % B1111 & CHKSUM_MASK
-    
-    
+serial_coms.establish_contact()
